requirements.txt
fastapi
uvicorn
streamlit
pandas
pytest
PyYAML
